Add some optimizations to plugin painting.
The simplest one is to disable blending when the plugin is opaque. The more
complicated one is to bypass webkit painting the background of plugins when we
know the plugin to be always on top and also opaque. The always on top flag is
currently set by a new "Private2" API. Bypassing WebKit makes animations
faster.

This is a re-land of the previous patch with a trivial compilation fix. This
also adds a clip rect to the GetBitmap... function so we can properly handle
plugins in nested iframes with proper clipping.

BUG=none
TEST=none
Original review URL: http://codereview.chromium.org/3421030

git-svn-id: http://src.chromium.org/svn/trunk/src@60426 4ff67af0-8c30-449e-8e8b-ad334ec8d88c