Remove flakiness from url_info tests
This test was testing to be sure that timing was
correctly monitored during the lifetime of 
these UrlHostInfo instances.  The new code defensively
checks values, and when tehy appear to be out of
spec, it verifies that the test bot is indeed
running slowly, and exits.  This will usually
run the test as expected, but gracefully
exit when the bot is just too slow (or
has too much variance in execution speed)
to run the test carefully.

r=phajdan.jr
BUG=55169
Review URL: http://codereview.chromium.org/3431028

git-svn-id: http://src.chromium.org/svn/trunk/src@60962 4ff67af0-8c30-449e-8e8b-ad334ec8d88c