Render crash in FormManager::FindCachedFormElement()
To address the vulnerability of stale WebFrame pointers in the FormManager's cache this CL changes the cache from a map (with the WebFrame pointer as "key") to a flat vector of simplified "FormElement*" items.

To avoid leaking memory, we need to still observe |frameDetached|, and use that as a signal to reap any associated WebFormElements or WebFormControlElements.

BUG=48857
TEST=FormMananagerTest.*, and manual test of regular form filling, form filling a form with sub-iframes, and form filling a form with sub-frames.

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=60949

Review URL: http://codereview.chromium.org/3492015

git-svn-id: http://src.chromium.org/svn/trunk/src@60999 4ff67af0-8c30-449e-8e8b-ad334ec8d88c