Enable the invalidating Core Animation plugin drawing model on 10.5
This also adds a layer of abstraction between the concept of the accelerated surface, and our cross-process implementation (which is something we needed to fix in general, but was necessary here to avoid bleeding TransportDIB references into code that's not supposed to know about process separation). There's no in-process implementation since we don't support in-process plugins on the Mac, but the abstraction is now there if we want to add one in the future.

BUG=32012
TEST=Plugins using invalidating Core Animation model should run on 10.5

Review URL: http://codereview.chromium.org/3449023

git-svn-id: http://src.chromium.org/svn/trunk/src@61199 4ff67af0-8c30-449e-8e8b-ad334ec8d88c