This is a monster CL.
It started as an attempt to put the decoder onto another thread.  However, this became complicated due to multiple object ownership transfers and coupling between the decode layer and the network layer; the decoder's states were highly coupled with how the network packets were processed.

This could probably be broken up slightly, but at this point, it's easier to just commit as a whole  The refactor includes:
  1) Making the decoder interface unaware of "network packet" types.
  2) Making the network layer process packets in order.
  3) Threading through asynchronous APIs all over the place.
  4) Simplifying the rectangle update protocol.
  5) Cleaning up object lifetime and ownership semantics between the decode layer and the renderer.

As of right now, the Verbatim format is still broken on the encode side because it uses the old protocol.

BUG=52883, 57351
TEST=still connects to chromoting_simple_host

Review URL: http://codereview.chromium.org/3305001

git-svn-id: http://src.chromium.org/svn/trunk/src@61402 4ff67af0-8c30-449e-8e8b-ad334ec8d88c