Fix browser crash when PageAction receives a cached image from ImageLoadingTracker.
Here is what is happening: 

A notification comes in, saying: Extension loaded.
This triggers a RefreshPageActionViews, which results in the vector of PageActionImageView objects being recreated.
While constructing one of those PageActionImageView objects, if the ImageLoadingTracker returns the results synchronously (because the image has been cached), we'll get OnImageLoaded during construction, which will call UpdatePageActions, starting the process of RefreshPageActionViews all over again (this time with a vector of PageActions that is still being constructed, and partially populated with null values).

The fix is to not call UpdatePageActions unless we have been assigned a parent, which avoids this re-entrancy during construction.

BUG=57333
TEST=New test: ExtensionApiTest.TestCrash57333

Review URL: http://codereview.chromium.org/3621007

git-svn-id: http://src.chromium.org/svn/trunk/src@61625 4ff67af0-8c30-449e-8e8b-ad334ec8d88c