Add placeholder text to DOM UI content settings exceptions <input>.
The actual placeholder attribute doesn't work because it disappears when the <input> is focused, and the <input> is focused as soon as it appears, so the text is placed as the contents of the <input>. Adding a new row should show [*.]example.com, and it should start as selected. Pressing enter in this one case should delete the row rather than inserting a [*.]example.com exception. Pressing escape in all cases should revert the row (delete it for Add'd rows).

BUG=35087, 57458
TEST=manual

Review URL: http://codereview.chromium.org/3573009

git-svn-id: http://src.chromium.org/svn/trunk/src@61860 4ff67af0-8c30-449e-8e8b-ad334ec8d88c