The IPC channel names and event names used by the service process now append the version string so that a browser with the same version only can talk to it.
In addition, the service process writes its version into shared memory (the name of this shared mem does not use a version) and also listens on a shutdown event. This can be used by a newer service process to shutdown an older version when it starts.

BUG=None
TEST=Unittests, with an existing service process running, let chrome autoupdate and then run the service process again. It should kill the old one.

Review URL: http://codereview.chromium.org/3603016

git-svn-id: http://src.chromium.org/svn/trunk/src@62066 4ff67af0-8c30-449e-8e8b-ad334ec8d88c