Record startup time closer to where the browser really started
startup time is used as a timestamp to discover new crashes (determined by
entries in crash dir since the recorded startup time) in the ui_test
framework. Initialing it in the constructor is problematic for PyAuto since
python unittest initialzes testcase objects for all tests together, so all
instances will have about the same startup time, and this leads to crashes
from one test being mismatched to the error reporting in another.

This CL fixes it by moving startup time initialization to SetUp, which is
closer to the time the browser actually gets launched.
Consciously, Not moving to LaunchBrowserAndServer.

BUG=51463

TEST=

Review URL: http://codereview.chromium.org/3670004

git-svn-id: http://src.chromium.org/svn/trunk/src@62314 4ff67af0-8c30-449e-8e8b-ad334ec8d88c