Compile webkit gpu code in an upstream build too.
I had originally disabled this when setting up the upstream
webkit.org build.  At the time we didn't compile any of src/gpu,
but now we do, so it's safe to enable this.

If we don't compile this, shared library (.so) builds in a
webkit only tree fail to link because of missing symbols.

Review URL: http://codereview.chromium.org/3722002

git-svn-id: http://src.chromium.org/svn/trunk/src@62330 4ff67af0-8c30-449e-8e8b-ad334ec8d88c