Adding a method to the Sync test harness that disables sync for a client.
The sync integration test harness does not provide tests with a way of
disabling a sync client, thereby blocking the automation of some tests.

This patch adds a method called DisableSyncForAllDatatypes() to the
ProfileSyncServiceTestHarness class. It also contains a refactor of some
of the inner methods, to make them support the SYNC_DISABLED state.

BUG=56297
TEST=Disable sync, make a local change, make a change in a partner, re-enable sync, etc.

Review URL: http://codereview.chromium.org/3647006

git-svn-id: http://src.chromium.org/svn/trunk/src@62441 4ff67af0-8c30-449e-8e8b-ad334ec8d88c