about:flags: Fix disabling experiments.
The problem was that the restart code would copy the switches added by about:flags to the restarted instance, so that e.g. --enable-expose-for-tabs would be passed on the the command line even though the experiment was disabled. Now the flags added by about:switches are explicitly removed before the restart.

Also wrap the switches added by about:flags with --start-flags-switches and --end-flags-switches switches. These unknown switches are ignored, but are useful to see on about:version.

Also add a bunch of unrelated unit tests.

BUG=56314
TEST=See bug.

Review URL: http://codereview.chromium.org/3813007

git-svn-id: http://src.chromium.org/svn/trunk/src@62777 4ff67af0-8c30-449e-8e8b-ad334ec8d88c