FBTF: Move BlobData ParamTrait structs from common_param_traits to render_messages.
The linking failures in chrome frame are due to common_param_traits adding a dependency on the webkit/ directory, which is compiled and used by nacl and chrome_frame for a minified chrome-common. render_messages, the only user of the BlobData param traits, is compiled in the full version of chrome/common/ and is where it needs to be placed.

Long term something needs to be done, but for now I've added a big warning message to common_param_traits.cc. (The webkit/ includes in the current common_param_traits only refer to values in headers currently, but this still should be resolved.)

BUG=none
TEST=compiles

Review URL: http://codereview.chromium.org/3614005

git-svn-id: http://src.chromium.org/svn/trunk/src@63263 4ff67af0-8c30-449e-8e8b-ad334ec8d88c