Autofilled values should match previewed values for previously autofilled forms.
It is possible to autofill a form more than once, e.g. by autofilling it and then clearing some of the fields.  We used to show different values for previewing vs. filling the form.  This patch makes those values consistent.

The behavior is far from perfect for yucky edge cases, but the whole approach to autfilling a single form multiple times is being rethought, so this should be fine in the interim.

BUG=60399
TEST=Manually inspected (again, this interaction is being reconsidered, so not going to add temporary tests).

Review URL: http://codereview.chromium.org/4030003

git-svn-id: http://src.chromium.org/svn/trunk/src@63800 4ff67af0-8c30-449e-8e8b-ad334ec8d88c