chromeos: Avoid assert if WM tells us to select invalid tab.
The window manager gets notified asynchronously when the
number of tabs in a window changes, so there's no guarantee
that the select-tab messages that the WM sends will always
have a valid index in them.  This change makes us ignore
invalid messages instead of CHECK()-ing when we receive
them.

BUG=chromium-os:8103
TEST=built an evil window manager that always sends index -1 and checked that chrome stops crashing with this change

Review URL: http://codereview.chromium.org/4089002

git-svn-id: http://src.chromium.org/svn/trunk/src@63803 4ff67af0-8c30-449e-8e8b-ad334ec8d88c