Reduce Jankometer measurements by only periodically sampling
The Jankometer is monitoring durations for each
task, but most of the time, this work is not used.
I changed the system to only sample the durations,
unless someone turns on a watchdog, and then we
monitor every event.  The sampling is deterministic,
(every N events is sampled) but should be fine 
when we aggregate across thousands of users in
a histogram.

r=brettw
Review URL: http://codereview.chromium.org/4279003

git-svn-id: http://src.chromium.org/svn/trunk/src@64991 4ff67af0-8c30-449e-8e8b-ad334ec8d88c