Added GPUBrowserTest.BrowserTestLaunchedWithOSMesa.
This just verifies that the OpenGL backend is OSMesa and does a simple smoke test that it is functional.

It is the browser test equivalent of GPUUITest.UITestLaunchedWithOSMesa. OSMesa should be required for both kinds of tests.

I moved both tests into the new chrome/test/gpu directory.

There are ongoing stability issues with OSMesa (and GL in general) on bots.

TEST=passes locally, try
BUG=none

Review URL: http://codereview.chromium.org/4424001

git-svn-id: http://src.chromium.org/svn/trunk/src@65126 4ff67af0-8c30-449e-8e8b-ad334ec8d88c