Try to improve message_loop efficiency when backlogged
The call to Now() uses a lock to handle roll over events, and that
might make the call more expensive in some scenarios.

Reduce the number of times we need to call Time::Now() when processing
timer events by processing all (fully ripe, and ready to run) events
with only one call to Now(). This should make the queue run faster as
we fall further behind.

I was hoping to see some perf improvement, but I think this will only
have an impact when we are overloaded, which problably doesn't happen
often on the bots.  I did some local histogramming, and I do see cases
where we get away with processing several (2-3) events with one call
to Now(), but that is apparently not enough to move the needle on the
perf bots.

r=darin
Review URL: http://codereview.chromium.org/4247003

git-svn-id: http://src.chromium.org/svn/trunk/src@65131 4ff67af0-8c30-449e-8e8b-ad334ec8d88c