Close URL requests associated with an instance when that instance goes away.
Not doing this causes a bug in shutdown (wehn we can't do fast shutdown and just kill the renderer).

The problem is that if a URLLoader is open (like maybe it's leaked by the plugin), the ResourceTracker will be the one to delete it. The ResourceTracker is a singleton which is deleted in the AtExit manager. If a URLLoader is deleted from there we'll crash trying to delete the corresponding WebKit loader which no longer exists (because the page associated with it went away).

BUG=61795
TEST=none
Review URL: http://codereview.chromium.org/4690003

git-svn-id: http://src.chromium.org/svn/trunk/src@65501 4ff67af0-8c30-449e-8e8b-ad334ec8d88c