Provide sync integration tests with a way to set a favicon for a bookmark.
The sync integration test framework does not provide tests with a way to
set a favicon for a bookmark and verify that it gets synced.

This patch adds a method called SetFavicon, that allows tests to set
either a generic favicon, or a variable favicon (using an index) for a
bookmark.

BUG=57234
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/4749003

git-svn-id: http://src.chromium.org/svn/trunk/src@65840 4ff67af0-8c30-449e-8e8b-ad334ec8d88c