Improve quality of DCHECK for IsOnIoThread
The old system was adaptive, and this new system
uses centralized constant to perform the assertion.

I also edited several comments, and added a cleaner reset
to the state of the *_sequence_number_ after we've serviced
all requests (resetting them back to -2, which was what
we had after we initialized).

BUG=62810
r=nkostylev,davemoore
Review URL: http://codereview.chromium.org/4688009

git-svn-id: http://src.chromium.org/svn/trunk/src@65955 4ff67af0-8c30-449e-8e8b-ad334ec8d88c