Fix some bugs in paint manager. Some of the class members were not getting initialized in the constructor.
Also, when a manual callback was executed, if a regular fluch happened to come
in the intervening time, we would cause an assertion failure because it would
always try to paint. Now when we get the callback, we can ignore it if there is
already a flush pending.

Original review=http://codereview.chromium.org/4886004/

git-svn-id: http://src.chromium.org/svn/trunk/src@66082 4ff67af0-8c30-449e-8e8b-ad334ec8d88c