[Chrome OS] Cleanup and modify Chrome-for-ChromeOS autotest wrappers
Some deps in the chrome tree and support classes on the Chrome OS side
have changed, so update our autotest wrappers to work again.  The new
code also takes advantage of some new features in the ChromeTestBase class.

I've also split the control files to each run 1/4 of the browser tests and ui tests.  I tried to come up with a way to divide them more cleanly, but this seems the best way to split them across multiple VMs with the infrastructure on the bots.  If we used autoserv to run smoke suite, we could be more clever here, but that's a big hammer for this problem.

BUG=chromium-os:8544
TEST=install chrome with tests, then run desktopui_BrowserTest and friends.  They should be able to run, though not necessarily succeed.

Review URL: http://codereview.chromium.org/4698004

git-svn-id: http://src.chromium.org/svn/trunk/src@66168 4ff67af0-8c30-449e-8e8b-ad334ec8d88c