The UI language rather than the locale is now used to pick Chrome's language on Windows.  (Also fixed some unrelated lint errors.)  With this change, l10n_util::GetApplicationLocale first checks for an override of the "configured locale" (in base::i18n) containing the list of preferred Windows UI languages.  The browser triggers an override early in startup before the ApplicationLocale is determined. In effect, the browser no longer uses ICU on Windows for language detection.  (This locale override mechanism is borrowed from the OS X port.)
Changes in Chrome Frame are largely a refactor, as some Win32 code in there has been moved into base/win.

Also cleaned up language selection in installer_util so that the proper language is chosen for the EULA, installer messages, and shortcuts.  In so doing, replaced hand-crafted lists of supported languages with either auto-generated lists (static consts) or logic so that the addition of translations in the future doesn't require code motion (that being said, there may be reason to update the alias and/or wildcard tables in language_selector.cc).  In so doing, this change unlocks Amharic, Farsi, and Swahili translations for installer messages and shortcuts.

BUG=39986,40496,26470
TEST=New MUI/Win32 calls are tested in base/win/i18n_unittest.cc.  To test the overall functionality, uninstall Chrome, remove intl.app_locale user pref, switch to a supported display language (via the "Keyboards and Languages" tab of Win7's "Regional and Language" control panel, and install with { "distribution": { "require_eula": true } } in master_preferences (via -installerdata arg to setup.exe).  If all goes well, both EULA and outer frame are in the same language as Windows.  Also, from gwilson: "Install system-level Chrome in audit mode on a new machine, then go through the out-of-box-experience, select a language, and the in -product EULA (triggered by "require_eula" : true) and Chrome's UI should be in the language that the user selected."

Review URL: http://codereview.chromium.org/4139010

git-svn-id: http://src.chromium.org/svn/trunk/src@66275 4ff67af0-8c30-449e-8e8b-ad334ec8d88c