chromeos: Add screen locker to window group *after* grab.
Otherwise, if it's already in its own window group, it won't
be able to break other widgets' grabs.

This also separates out the showing and grabbing code, so we
can show the widget before mapping the locker window and then
grab the pointer and keyboard afterwards.

BUG=chromium-os:8999,chromium-os:9200
TEST=opened a context menu and checked that screen locks immediately.  also tested that the original problem with modal dialogs didn't regress

Review URL: http://codereview.chromium.org/5013003

git-svn-id: http://src.chromium.org/svn/trunk/src@66287 4ff67af0-8c30-449e-8e8b-ad334ec8d88c