Cleaned up thread safety for GpuProcessHost and GpuProcessHostUIShim.
Guarded static members with an assertion that they are invoked on the right thread.

Guarded non-static members with an assertion that NonThreadSafe::CalledOnValidThread() is true.

GpuProcessHost automatically routes unhandled control messages to the UI shim.

Moved gpu_info to the UI shim, which is the thread it is accessed on.

TEST=webgl, CSS 3D locally, try
BUG=none

Review URL: http://codereview.chromium.org/5130001

git-svn-id: http://src.chromium.org/svn/trunk/src@66482 4ff67af0-8c30-449e-8e8b-ad334ec8d88c