Add shared macros for stringizing and converting ANSI string constants (in particular defined by macros) to wide string constants.
Convert existing locally-defined stringizing to use the shared macros.

Unit tests for the shared macros.

This also fixes a minor bug in ceee_module_util.cc where I
accidentally quoted a string constant I only meant to convert to wide
(this caused no bug, but was unintended, so the change in semantics
in that file in the current change is intentional).

BUG=none
TEST=automated tests


Review URL: http://codereview.chromium.org/5103001

git-svn-id: http://src.chromium.org/svn/trunk/src@66579 4ff67af0-8c30-449e-8e8b-ad334ec8d88c