Mac: Only clear the background of CoreAnimation plugins if we're going to draw into them.
Previously, the logic was:

1.) Clear plugin background
2.) If the plugin didn't update, return early
3.) Paint plugin
4.) "Swap buffers"

But the "Swap buffers" step only unbound and rebound an FBO object If the plugin didn't change, its backing store would contain transparent black, and if the graphics driver decided to flush the FBO for another reason than a "swap buffers" call, the blackness would show up in the browser.

This CL swaps steps 1 and 2, so even if the FBO is flushed for some unrelated reason, we display something valid.

BUG=60341
TEST=Open the file attached to the bug. Resize the window for a few minutes, put the computer to sleep and back on, resize window for a few more minutes. The plugin area (the 191x60px rect in the middle) shouldn't become black. YouTube should still work. CPU usage shouldn't be worse than it was before for the browser, plugin, and renderer processes.

Review URL: http://codereview.chromium.org/5220002

git-svn-id: http://src.chromium.org/svn/trunk/src@66631 4ff67af0-8c30-449e-8e8b-ad334ec8d88c