Fix a couple of locations that should have been passing show_window = true in NavigateParams.
BUG=none
TEST=

Bug 1: open an incognito window. switch back to the non-incognito window then right click a link and choose "open in incognito window." link should be opened in new tab in existing incognito window and that window activated. bug is that this happens but window is not activated.
Bug 2: open an app frame (e.g. chrome --app=http://mail.google.com/). click on a link that would open a new tab (which can't be opened in an app frame). If an existing chrome window would be the target that window is never activated.
Review URL: http://codereview.chromium.org/5219005

git-svn-id: http://src.chromium.org/svn/trunk/src@66818 4ff67af0-8c30-449e-8e8b-ad334ec8d88c