Hide ChromeDnsCertProvenanceChecker behind a factory.
Directly calling NSS code from a .cc file which is linked into both
Chrome and Chrome Frame causes a bad reference to libnspr.dll to be
emitted on Windows. wtc informs me that the `solution' to this is to
break the code up so that there's no direct chain from common code to
NSS code.

This change puts ChromeDnsCertProvenanceChecker into its own file and
adds a factory for it. The factory is modeled after the SSLClientSocket
factory in net/socket (which was added for the some of the same
reasons).

BUG=none
TEST=compiles on Chrome Frame builders.

http://codereview.chromium.org/5260001

git-svn-id: http://src.chromium.org/svn/trunk/src@67098 4ff67af0-8c30-449e-8e8b-ad334ec8d88c