Initialize DOMUI handlers if the DOM itself is re-used
Back/forward causes a page navigation, triggering a new DOMUI to be created. Since the page is the same, however, WebKit doesn't actually do a navigation, and the DOM remains intact, so the onload handler doesn't fire and cause the newly-created DOMUI to be initialized.

This change notifies the DOMUI object when it has been bound to a RenderView that may have an existing DOM on the HTML/JS side, so that it can tell the page to re-trigger initialization if necessary.

BUG=60768
TEST=Open DOMUI prefs, go to another pref section, press the Back button, then try to change the search engines.

Review URL: http://codereview.chromium.org/5120005

git-svn-id: http://src.chromium.org/svn/trunk/src@67611 4ff67af0-8c30-449e-8e8b-ad334ec8d88c