Benchmark tool for GPU-accelerated video rendering
This is a benchmarking program that times how fast a shader
can color-convert and render video frames to the screen. The
program takes a file with raw YUV frames, the number of
frames in the file, and the dimensions of the video file as
parameters and outputs the max frames per second obtained
when rendering the video using various painting/shading
techniques.

BUG=48633
TEST=compiles

Review URL: http://codereview.chromium.org/4873002

git-svn-id: http://src.chromium.org/svn/trunk/src@67887 4ff67af0-8c30-449e-8e8b-ad334ec8d88c