Change HttpProxyClientSocketPool so that when a new SpdyStream is created, it uses the NetLog from the SpdySession, not from the ConnectJob, so that we avoid circular dependencies.
I attempted to verify the buggy behavior and the fix by adding a
DCHECK() to ConnectJob::set_socket (which is the source of the 
cycle you detected), but unfortunately, it appears that our unit
tests seems to use the same NetLog everywhere.  So to test this,
I added a very specific DCHECK in passive_log_collector.cc
which failed on every HTTPS request through a SPDY proxy when
using the old code, and which never failed using this new code.


BUG=65217
TEST=none

Review URL: http://codereview.chromium.org/5587005

git-svn-id: http://src.chromium.org/svn/trunk/src@68209 4ff67af0-8c30-449e-8e8b-ad334ec8d88c