Provide a safety net for Windows hooks.
As mentioned in http://crbug.com/65655, we have seen cases where the IE process would crash in a Windows Hook called after static uninitialization. We are guessing that this occurs when the thread running the CEEE BHO got terminated without tearing down the BHO and thus we never unhooked.

This CL adds a module level set of hooks to be unhooked when the module gets terminated (before the static uninitialization).

BUG=65655
TEST=None

Review URL: http://codereview.chromium.org/5544014

git-svn-id: http://src.chromium.org/svn/trunk/src@68552 4ff67af0-8c30-449e-8e8b-ad334ec8d88c