Teach PPAPI proxy to share image memory on Mac OS X
Instead of just using base::SharedMemory, we use SysV shared memory on Linux so
that we can share memory with X.  Normally we abstract this detail away with
TransportDIB, but the TransportDIB lives in chrome/app, which is outside the
dependency cone of ppapi/proxy.

This patch creates a new abstraction at this layer of the dependency graph
called ImageData::ImageHandle that represents shared memory for the purpose of
image data.  This patch also fills in the implementation of this abstraction on
Mac OS X.  Consequently, out-of-process PPAPI plugins on Mac can now draw to
the screen.

Review URL: http://codereview.chromium.org/5658003

git-svn-id: http://src.chromium.org/svn/trunk/src@68569 4ff67af0-8c30-449e-8e8b-ad334ec8d88c