Fix browser crash in ExtensionPrefs.
I am unable to repro this, despite numerous attempts, but from looking
at the crash dump it seems that we get a valid ExtensionPrefs pointer
but a NULL DictionaryValue pointer. Looking at the other callers for
GetExtensionPref shows that they all expect a potentially NULL result,
for example in GetExtensionState (which even documents that it can be
NULL for --load-extension).

I think a NULL check here is therefore in order too.

BUG=64853
TEST=QA will need to test this manually because I can't get it to crash for me.

Review URL: http://codereview.chromium.org/5601005

git-svn-id: http://src.chromium.org/svn/trunk/src@68572 4ff67af0-8c30-449e-8e8b-ad334ec8d88c