Re-land 69416 - IndexedDB: Preparatory patch for using Date objects as keys.
It was reverted in 69437 because it seemed to have broken

SSLUITest.TestHTTPWithBrokenHTTPSResource
SSLUITest.TestHTTPSExpiredCertAndProceed
SSLUITest.TestCNInvalidStickiness
SSLUITest.TestRefNavigation
on Linux & Linux64.

But those tests passed on the try servers, still pass on my machine, and have nothing to do with my patch (that I can imagine).

Landing when the tree is calm to see how it goes.

BUG=64060
TEST=none

TBR=siggi@chromium.org
Review URL: http://codereview.chromium.org/5983001

git-svn-id: http://src.chromium.org/svn/trunk/src@69526 4ff67af0-8c30-449e-8e8b-ad334ec8d88c