Clobber the state of the next-protocol-negotiation when disabling spdy.
We basically have two sets of config for SPDY - one is a general enablement
flag, the other is the next-protocol advertisement.  We previously could
configure to advertise SPDY over NPN, but have SPDY disabled; this would lead
to the bug with broken 'download' when accessing a SPDY page.

BUG=66891
TEST=none



Review URL: http://codereview.chromium.org/6007002

git-svn-id: http://src.chromium.org/svn/trunk/src@69891 4ff67af0-8c30-449e-8e8b-ad334ec8d88c