Defines SSLServerSocket and implements SSLServerSocketNSS
Defines a SSLServerSocket interface. Implement this interface using NSS as
SSLServerSocketNSS.

This is the first version of the code. It disables several functions of NSS
like caching, session ticket, reneogotiation, etc. This is implemented to
suit the needs of Chromoting. Additional features of this socket will be added
when necessary.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/5746003

git-svn-id: http://src.chromium.org/svn/trunk/src@70041 4ff67af0-8c30-449e-8e8b-ad334ec8d88c