Change extension unload notification to indicate updates.
When an extension in being unloaded, some listeners want to know if it's 
because the extension is being updated to a newer version, or disabled. This
changes the details sent to include a reason.

Also this removes the EXTENSION_UNLOADED_DISABLED notification, since only a
small number of places actually cared about the disctinction between 
EXTENSION_UNLOADED and EXTENSION_UNLOADED_DISABLED, and puts that information
into the details as well.

BUG=65510
TEST=Should be covered by existing unit & browser tests.


Review URL: http://codereview.chromium.org/5968009

git-svn-id: http://src.chromium.org/svn/trunk/src@70104 4ff67af0-8c30-449e-8e8b-ad334ec8d88c