Move app/win_util to app/win and fix the namespace usage. Split out the two classes: ScopedComInitializer and ScopedCOMem (which I renamed) to separate files.
I removed the win_util_path file which had one function in it and moved the function to win_util.
Somehow, this was getting picked up by the nacl64 build and the call in sandbox_policy was then not being
defined. I just implemented the function in-plcae since it's just a simple wrapper around a Windows API call.

TEST=it compiles
BUG=none
Review URL: http://codereview.chromium.org/6013009

git-svn-id: http://src.chromium.org/svn/trunk/src@70343 4ff67af0-8c30-449e-8e8b-ad334ec8d88c