Re-use existing out-of-process plugins if one is already loaded. Otherwise sites with multiple plugins will load that many child processes, which is bad.
This works by having creators of out-of-process plugins add it to a list in the
PepperPluginRegistry. The items are deleted from this list by a new delegate
implementation for the module to tell the registry that it's being destroyed.

It's a little awkward since the PluginModule can't know about the
PepperPluginRegistry, and all other pepper "delegates" are associated with the
Instance/RenderView, so by definition can't be used when the module is being
torn down since there are no instances.

It might have been nice to have an Add function on the lifetime delegate so
that the module will add itself to the list. But the plugin path name is
currently not available in all code paths, and it seemed messy to add.

TEST=manual
BUG=none
Review URL: http://codereview.chromium.org/6085007

git-svn-id: http://src.chromium.org/svn/trunk/src@70457 4ff67af0-8c30-449e-8e8b-ad334ec8d88c