Popups opened with the noreferrer parameter within ChromeFrame should open in the host browser. This support already exists in Chrome where the renderer routes the url back to the browser and eventually to the TabContentsDelegate.
However in ChromeFrame if the current window is a popup window then attempting to open this URL
in a tab fails on IE7 and IE8 as popups don't support this functionality. Workaround as per
msdn documentation is to reissue the navigation request with the navOpenInNewWindow flag.

Fixed a bug in the ChromeFrame tests IEEventSink::GetRendererWindow function where at times
it would fail to return the correct HWND in the window hierarchy when the page was rendered
in ChromeFrame.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=68183

BUG=68183
TEST=Covered by new ChromeFrame test.

Review URL: http://codereview.chromium.org/6002017

git-svn-id: http://src.chromium.org/svn/trunk/src@70676 4ff67af0-8c30-449e-8e8b-ad334ec8d88c