Add an instance parameter to var objects, audio, and the 2D API. This replaces the module in most cases.
This will be used in the proxy to multiplex one plugin across multiple renderer processes. We need the instance in the proxy to know which process to send it to.

I added a deprecated var object creation function for native client, which
depends on the module and this is very difficult to change. Because it doesn't
have the multiplexing requirements, this is fine for now.

TEST=ppapi ui tests
BUG=none
Review URL: http://codereview.chromium.org/6085009

git-svn-id: http://src.chromium.org/svn/trunk/src@70721 4ff67af0-8c30-449e-8e8b-ad334ec8d88c