Move the in-process WebGL implementation from the WebKit to the Chromium repository to break circular dependencies on Chromium's OpenGL bindings in app/gfx/gl.
The minimal changes to the code were made in order to break
dependencies on WebCore types and to make the code fit the Google C++
style guide as closely as possible. There are many references to
undesirable types like "long" in the WebGraphicsContext3D interface,
which are the source of the majority of lint errors and which will be
fixed in follow-on work.

The upstream WebGraphicsContext3DDefaultImpl will be deleted after
this lands.

Tested:
 - Various WebGL demos in Chrome
 - Built and ran WebGL layout tests in DRT

BUG=none
TEST=none (see above)

Review URL: http://codereview.chromium.org/6100009

git-svn-id: http://src.chromium.org/svn/trunk/src@71069 4ff67af0-8c30-449e-8e8b-ad334ec8d88c