Mac: Explicitly set the colorspace on SkBitmap -> CGImageRef conversions.
The color space was hardcoded as "generic rgb" in skia. This is not always correct, also skia is changing this color space around a lot currently. To protect us from their unreliable default, hardcode "generic rgb" as default on our side for now, but make it possible for clients to provide their own color space.

Use this to let tabpose and the favicon code pass in the device colorspace.

BUG=24267,50307
TEST=Open tabpose. Delayed thumbnails should look like backing-store backed thumbnails. The colors of favicons should now match other browsers.

Review URL: http://codereview.chromium.org/6117006

git-svn-id: http://src.chromium.org/svn/trunk/src@71208 4ff67af0-8c30-449e-8e8b-ad334ec8d88c