Use a random number to generate a self-signed certiciate for chromoting
Chromoting host to use a random number to generate self-signed certificate.
This method is still not optimal, in case of collision SSLServerSocketNSS will
fail to start due to NSS rejecting the certificate.

Using a random number is a temporary solution until we can assign certificates
to the host.

BUG=70013
TEST=Connect to the same chromoting host after client is disconnected.

Review URL: http://codereview.chromium.org/6315009

git-svn-id: http://src.chromium.org/svn/trunk/src@72745 4ff67af0-8c30-449e-8e8b-ad334ec8d88c