BMM: Handle middle click correctly.
This fixes a regression. Originally we listened to click events for middle click. The WebKit was change to not fire click events for middle clicks so I added code to work around that. The WebKit change caused a lot of regressions and was reverted. Now the BMM opened 2 background tabs (one for click and one for the work around). With this change I no longer listen to click and only do the work around.

BUG=70929
TEST=Open bookmark manager.
     Middle click on a bookmark.
     Only one background tab should be opened.

Review URL: http://codereview.chromium.org/6332020

git-svn-id: http://src.chromium.org/svn/trunk/src@72843 4ff67af0-8c30-449e-8e8b-ad334ec8d88c