Revert 72877 - Make BookmarkModelVerifier wait for favicons to load before extracting them
Reason for revert: Favicon sync test is still flaky on the buildbots.

The test case for bookmark favicon sync was failing in its verification
step because GetFavicon sometimes returned a blank favicon for a URL.
This was because GetFavicon can be asynchronous if the favicon of a
bookmark node is not already loaded.

This patch modifies the method that compares favicons, and takes care of
cases where the favicon has not yet loaded by adding a conditional wait.

BUG=69694
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6347016

TBR=rsimha@chromium.org
Review URL: http://codereview.chromium.org/6359017

git-svn-id: http://src.chromium.org/svn/trunk/src@72910 4ff67af0-8c30-449e-8e8b-ad334ec8d88c