Fix data type configuration race by making DataTypeManagerImpl::DownloadReady() non-reentrant
SyncBackendHost::ConfigureDataTypes() has a DCHECK in it that makes sure that only one sync data type configure task can run at a time.

The sync integration test setup flow involves 2 indirect calls to SyncBackendHost::ConfigureDataTypes():

1) ProfileSyncServiceHarness::SetupSync() starts by calling ProfileSyncService::OnUserChoseDatatypes(), that indirectly calls SyncBackendHost::ConfigureDataTypes() and schedules a configure task.
2) ProfileSyncServiceHarness::SetupSync() then runs the message loop in order to wait for observed_passphrase_required(). This sets off a series of events that lead calls to SyncBackendHost::Core::NotifyResumed() and DataTypeManagerImpl::Restart(), that in turn calls SyncBackendHost::ConfigureDataTypes().

It's possible at times that the configure task queued in #1 has not yet been serviced when #2 is being scheduled, causing the DCHECK to fire.

This patch modifies DataTypeManagerImpl::DownloadReady() so that it pushes a Restart() task to the message loop if one is already in progress so that the stack can be unwound, instead of directly calling Restart(), thereby eliminating the race.

BUG=70869, 42448
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6250074

git-svn-id: http://src.chromium.org/svn/trunk/src@73419 4ff67af0-8c30-449e-8e8b-ad334ec8d88c