Mac: Enable "Check Spelling While Typing" in Edit menu
Currently the "Edit -> Spelling and Grammar -> Check Spelling While Typing" menu item is always disabled.

The problem is that the logic to enable / disable this item lives in the render process and not in the UI.

This patch implements a generic message that the render process can send to the browser to update the state of view commands. This is used to enable and disable the "Check Spelling While Typing" menu item.

BUG=38440
TEST=Clicked in an edit box and verified that the "Check Spelling While Typing" menu item was enabled. Verified that changing the value from the context menu correctly updates the Edit menu as well.


git-svn-id: http://src.chromium.org/svn/trunk/src@73664 4ff67af0-8c30-449e-8e8b-ad334ec8d88c