Have a consistent UA string with chromeframe token
When using 'PostPlatform' registry, chromeframe UA token
appears in the middle whereas when we dynamically append
the UA we place it at the end. This has caused issues with
sites looking for it in a specific place. 

This change modifies the logic while dynamically appending
the UA to insert it consistently in the place where 
PostPlatform puts it.

BUG=70024	
TEST=covered by HtmlUtilUnittest.AddChromeFrameToUserAgentValue


Review URL: http://codereview.chromium.org/6246088

git-svn-id: http://src.chromium.org/svn/trunk/src@73830 4ff67af0-8c30-449e-8e8b-ad334ec8d88c