Attempt 3 at: Splits ChromeURLDataManager into 2 chunks: . ChromeURLDataManager is no longer a singleton and is always used on   the UI thread. ChromeURLDataManager is now profile specific (you get   from the profile). . ChromeURLDataManagerBackend handles the URLRequests and the   DataSources. ChromeURLDataManagerBackend is created by   ChromeURLRequestContext.
All DataSources are now profile specific. There were two that wanted
to be global, but have been converted.

This differs from the last version in that DataSource::SendResponse
does nothing if the DataSource is schedule for deletion. This is
necessary otherwise SendResponse will up the ref count and by the time
the request is processed on the IO thread chances are the DataSource
will have been deleted. And if it wasn't deleted, it'll get scheduled
for deletion again. <insert comment here about the perils of using
delayed deletion> This has always been possible, it just appears to be
more likely with my patch.

BUG=52022 71868
TEST=none

Review URL: http://codereview.chromium.org/6479007

git-svn-id: http://src.chromium.org/svn/trunk/src@74432 4ff67af0-8c30-449e-8e8b-ad334ec8d88c