Temporarily disable restrictions on who can request chrome-extension://.
These checks were causing lots of subtle bugs and fixing them will be difficult to get correct for the upcoming milestone. The checks were originally intended to be an 'extra line of defense' and are not critical for security.

I will reintroduce them separately on trunk.

BUG=72455,72407

Review URL: http://codereview.chromium.org/6478019

git-svn-id: http://src.chromium.org/svn/trunk/src@74511 4ff67af0-8c30-449e-8e8b-ad334ec8d88c