gtk: Improve fullscreen RenderWidgetHostViewGtk.
This was previously using the same code as for popup windows
(i.e. <select> and autocomplete), which creates an
override-redirect/popup window and grabs the pointer and
keyboard.  This is the wrong way to make a window
fullscreen: since the window is override-redirect, the
window manager will ignore the fullscreen request, and since
the input is grabbed, other X clients' key grabs (such as
Alt-Tab) won't work.

This change makes us instead open a regular toplevel window.

BUG=chromium-os:11545
TEST=checked that brightness and volume keys work while Pepper Flash is fullscreen and that <select> popups still work

Review URL: http://codereview.chromium.org/6462017

git-svn-id: http://src.chromium.org/svn/trunk/src@74534 4ff67af0-8c30-449e-8e8b-ad334ec8d88c