Add plumbing to webRequest API to support event filtering.
This introduces a new kind of event (WebRequestEvent) to handle the extra parameters to addListener. Behind the scenes, adding a listener creates a unique sub-event associated with that set of filter+extraInfo parameters. When we dispatch the event, we dispatch only those sub-events that match the required filters.

BUG=60101
TEST=no

Review URL: http://codereview.chromium.org/6250152

git-svn-id: http://src.chromium.org/svn/trunk/src@74887 4ff67af0-8c30-449e-8e8b-ad334ec8d88c