views transformation: First cut.
This is a first cut at the transformable views. Skia transformation matrices are used to apply transformation (rotate, scale, translation). Support for clipping is also added (but not through the matrix). Currently, the transformation is applied only for painting. A future CL will apply the transformation for various events.

BUG=none
TEST=ViewTest.TransformPaint

Review URL: http://codereview.chromium.org/6500008

git-svn-id: http://src.chromium.org/svn/trunk/src@75023 4ff67af0-8c30-449e-8e8b-ad334ec8d88c